+++ Emulator/Files/DiskFiles/DMSFile.h
@@ -50,7 +50,7 @@ public:
     isize numSectors() const override { return adf->numSectors(); }
     BootBlockType bootBlockType() const override { return adf->bootBlockType(); }
     const char *bootBlockName() const override { return adf->bootBlockName(); }
-    void readSector(u8 *target, long s) const override { return adf->readSector(target, s); }
-    void readSector(u8 *target, long t, long s) const override { return adf->readSector(target, t, s); }
+    void readSector(u8 *target, isize s) const override { return adf->readSector(target, s); }
+    void readSector(u8 *target, isize t, isize s) const override { return adf->readSector(target, t, s); }
     bool encodeDisk(class Disk *disk) override { return adf->encodeDisk(disk); }
 };
